eds-extension: don't overwrite calendar name with account name
Binding the account name to the source name causes the calendar names to
be overwritten with the account name.  Keeping the source name updated
makes sense only for the primary calendar associated with an account,
which indeed is named after the account.  In Ubports, the account name
never changes, so the binding is actually unnecessary; we still keep the
binding, out of completeness.

Fixes: https://github.com/ubports/ubuntu-touch/issues/762